Update CPE301 Final
This is the working code I have for the currently made circuit, still requires better commenting, water sensor, LCD output and stepper motor implementation. I will be adding the rest of the hardware tonight. It would be appreciated if someone could find the corresponding changes that need to be made to replace functions such as pinMode with register addressing